Add sign-in buttons to landing page
